requirements.txt
requests>=2.31
python-dotenv>=1.0
streamlit>=1.30
pandas>=2.1
plotly>=5.18
openpyxl>=3.1
pytest>=7.4
ruff>=0.1
